renamed to tool view
