+++ tutorials/eu.udig.tutorials.tool-view/plugin.xml
@@ -7,14 +7,14 @@
          point="org.eclipse.core.runtime.applications">
       <application>
          <run
-               class="eu.udig.tutorials.alertapp.Application">
+               class="eu.udig.tutorials.toolview.Application">
          </run>
       </application>
    </extension>
    <extension
          point="org.eclipse.ui.perspectives">
       <perspective
-            class="eu.udig.tutorials.alertapp.Perspective"
+            class="eu.udig.tutorials.toolview.Perspective"
             fixed="false"
             id="X.perspective"
             name="Perspective">
@@ -23,7 +23,7 @@
    <extension
          point="org.eclipse.ui.views">
       <view
-            class="eu.udig.tutorials.alertapp.View"
+            class="eu.udig.tutorials.toolview.View"
             icon="icons/view16/alert_view.png"
             id="X.view"
             name="View">
@@ -59,7 +59,7 @@
          point="net.refractions.udig.project.ui.tool">
       <modalTool
             categoryId="eu.udig.tutorials.alert-app.alerts"
-            class="eu.udig.tutorials.alertapp.SendAlertTool"
+            class="eu.udig.tutorials.toolview.SendAlertTool"
             icon="icons/etool16/alert_mode.png"
             id="eu.udig.tutorials.alert-app.sendalerttool"
             name="Send Alert"
@@ -80,7 +80,7 @@
    <extension
          point="net.refractions.udig.project.ui.toolManagers">
       <toolManager
-            class="eu.udig.tutorials.alertapp.CustomToolManager"
+            class="eu.udig.tutorials.toolview.CustomToolManager"
             id="eu.udig.tutorials.alert-app.customToolManager">
       </toolManager>
    </extension>
@@ -107,7 +107,7 @@
    <extension
          point="net.refractions.udig.mapgraphic.mapgraphic">
       <mapGraphic
-            class="eu.udig.tutorials.alertapp.ShowAlertsMapGraphic"
+            class="eu.udig.tutorials.toolview.ShowAlertsMapGraphic"
             icon="icons/view16/alert_view.png"
             id="eu.udig.tutorials.alert-app.showalertmapgraphic">
       </mapGraphic>
