+++ tutorials/eu.udig.tutorials.tool-view/src/eu/udig/tutorials/toolview/ApplicationActionBarAdvisor.java
@@ -1,4 +1,4 @@
-package eu.udig.tutorials.alertapp;
+package eu.udig.tutorials.toolview;
 
 import org.eclipse.ui.application.ActionBarAdvisor;
 import org.eclipse.ui.application.IActionBarConfigurer;
