+++ tutorials/eu.udig.tutorials.tool-view/src/eu/udig/tutorials/toolview/ApplicationWorkbenchAdvisor.java
@@ -1,4 +1,4 @@
-package eu.udig.tutorials.alertapp;
+package eu.udig.tutorials.toolview;
 
 import org.eclipse.ui.application.IWorkbenchWindowConfigurer;
 import org.eclipse.ui.application.WorkbenchAdvisor;
