+++ tutorials/eu.udig.tutorials.tool-view/src/eu/udig/tutorials/toolview/Perspective.java
@@ -1,4 +1,4 @@
-package eu.udig.tutorials.alertapp;
+package eu.udig.tutorials.toolview;
 
 import org.eclipse.ui.IPageLayout;
 import org.eclipse.ui.IPerspectiveFactory;
