+++ tutorials/eu.udig.tutorials.tool-view/src/eu/udig/tutorials/toolview/View.java
@@ -1,4 +1,4 @@
-package eu.udig.tutorials.alertapp;
+package eu.udig.tutorials.toolview;
 
 import java.io.File;
 import java.io.IOException;
